+++ packages/arcgis-rest-auth/src/UserSession.ts
@@ -109,7 +109,7 @@ export interface IOAuth2Options {
 
   /**
    * Duration (in minutes) that a token will be valid. Defaults to 20160 (two weeks).
-   * 
+   *
    * @deprecated use 'expiration' instead
    */
    duration?: number;
@@ -297,13 +297,14 @@ export class UserSession implements IAuthenticationManager {
 
   /**
    * Begins a new browser-based OAuth 2.0 sign in. If `options.popup` is `true` the
-   * authentication window will open in a new tab/window otherwise the user will
-   * be redirected to the authorization page in their current tab/window.
+   * authentication window will open in a new tab/window and the function will return
+   * Promise&lt;UserSession&gt;. Otherwise, the user will be redirected to the
+   * authorization page in their current tab/window and the function will return `undefined`.
    *
    * @browserOnly
    */
   /* istanbul ignore next */
-  public static beginOAuth2(options: IOAuth2Options, win: any = window) {
+  public static beginOAuth2(options: IOAuth2Options, win: any = window): Promise<UserSession> | undefined {
 
     if(options.duration) {
       console.log("DEPRECATED: 'duration' is deprecated - use 'expiration' instead");
@@ -387,7 +388,7 @@ export class UserSession implements IAuthenticationManager {
   }
 
   /**
-   * Completes a browser-based OAuth 2.0  in. If `options.popup` is `true` the user
+   * Completes a browser-based OAuth 2.0 sign in. If `options.popup` is `true` the user
    * will be returned to the previous window. Otherwise a new `UserSession`
    * will be returned. You must pass the same values for `options.popup` and
    * `options.portal` as you used in `beginOAuth2()`.
@@ -502,9 +503,9 @@ export class UserSession implements IAuthenticationManager {
     if (!win && window) {
       win = window;
     }
-    // Declar handler outside of promise scope so we can detach it
+    // Declare handler outside of promise scope so we can detach it
     let handler: (event: any) => void;
-    // return a promise that will resolve when the handler recieves
+    // return a promise that will resolve when the handler receives
     // session information from the correct origin
     return new Promise((resolve, reject) => {
       // create an event handler that just wraps the parentMessageHandler
@@ -570,7 +571,7 @@ export class UserSession implements IAuthenticationManager {
     const { portal, clientId, redirectUri, refreshTokenTTL }: IOAuth2Options = {
       ...{
         portal: "https://www.arcgis.com/sharing/rest",
-        refreshTokenTTL: 1440,
+        refreshTokenTTL: 20160,
       },
       ...options,
     };
@@ -591,7 +592,7 @@ export class UserSession implements IAuthenticationManager {
         refreshToken: response.refreshToken,
         refreshTokenTTL,
         refreshTokenExpires: new Date(
-          Date.now() + (refreshTokenTTL - 1) * 1000
+          Date.now() + (refreshTokenTTL - 1) * 60 * 1000
         ),
         token: response.token,
         tokenExpires: response.expires,
@@ -705,7 +706,7 @@ export class UserSession implements IAuthenticationManager {
   public readonly redirectUri: string;
 
   /**
-   * Duration of new OAuth 2.0 refresh token validity.
+   * Duration of new OAuth 2.0 refresh token validity (in minutes).
    */
   public readonly refreshTokenTTL: number;
 
@@ -780,7 +781,7 @@ export class UserSession implements IAuthenticationManager {
     this.provider = options.provider || "arcgis";
     this.tokenDuration = options.tokenDuration || 20160;
     this.redirectUri = options.redirectUri;
-    this.refreshTokenTTL = options.refreshTokenTTL || 1440;
+    this.refreshTokenTTL = options.refreshTokenTTL || 20160;
     this.server = options.server;
 
     this.federatedServers = {};
@@ -1028,7 +1029,7 @@ export class UserSession implements IAuthenticationManager {
     const [match, protocol, domainAndPath] = root.match(/(https?:\/\/)(.+)/);
     const [domain, ...path] = domainAndPath.split("/");
 
-    // only the domain is lowercased becasue in some cases an org id might be
+    // only the domain is lowercased because in some cases an org id might be
     // in the path which cannot be lowercased.
     return `${protocol}${domain.toLowerCase()}/${path.join("/")}`;
   }
@@ -1067,7 +1068,7 @@ export class UserSession implements IAuthenticationManager {
     return (event: any) => {
       // Verify that the origin is valid
       // Note: do not use regex's here. validOrigins is an array so we're checking that the event's origin
-      // is in the array via exact match. More info about avoiding postMessave xss issues here
+      // is in the array via exact match. More info about avoiding postMessage xss issues here
       // https://jlajara.gitlab.io/web/2020/07/17/Dom_XSS_PostMessage_2.html#tipsbypasses-in-postmessage-vulnerabilities
       const isValidOrigin = validOrigins.indexOf(event.origin) > -1;
       // JSAPI handles this slightly differently - instead of checking a list, it will respond if