+++ packages/arcgis-rest-feature-layer/src/index.ts
@@ -6,6 +6,7 @@ export * from "./add";
 export * from "./update";
 export * from "./delete";
 export * from "./applyEdits";
+export * from "./getAllLayersAndTables";
 export * from "./getAttachments";
 export * from "./addAttachment";
 export * from "./updateAttachment";