+++ CHANGELOG.md
@@ -5,7 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-### Unrealased
+### Unreleased
+## [3.2.1](https://github.com/Esri/arcgis-rest-js/compare/v3.2.0...v3.2.1) - July 9th 2021
+
+### Fixed
+
+- **arcgis-rest-feature-layer** export `getAllLayersAndTables`
+
+## [3.2.0](https://github.com/Esri/arcgis-rest-js/compare/v3.1.2...v3.2.0) - July 9th 2021
+
+### Added
+
+- **arcgis-rest-feature-layer** add `getAllLayersAndTables` and `parseServiceUrl`
 
 ## [3.1.2](https://github.com/Esri/arcgis-rest-js/compare/v3.1.1...v3.1.2) - June 16th 2021